libs/skills_agent/pyproject.toml
[project]
name = "skills-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.9",
    "langchain-anthropic>=1.3.2",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.8",
    "pyyaml>=6.0.3",
]
